Add symmetric generalized eigenvalue problems.
Need to change design strategy so that these functions do minimal amount of copying.  For the most part, it has turned out that I build a matrix and use it only once for a call, rather than repeatedly.  Since repeated use is the exception rather than the norm, the code should assume single use and expect the client to preserve matrices when necessary.

"const" on a parameter is not sufficient indication that the value will be preserved, because sometimes it is useful to allow an automatic conversion.  Instead, the doc comments should explicitly state which paramters are preserved, and the client should expect all others to be modified.