More rigorous naming:  Renamed YVYUChar->UYVYChar and VYUYChar->YUYVChar to follow memory order convention.  Changed implementation of RGBAChar to follow memory order convention.
Changed get/setRGBA() to guarantee an order within machine word (which is of course invariant to endianness).

With the above two changes and a few other small tweaks to the code, it should now work for big-endian as well as small-endian, so removed endian warning.  However, this hasn't been tested on an actual big-endian yet.

Added RGBChar.  This is the format used by all the image i/o libraries that I currently link to.

Still to do:  make RGBAChar and RGBChar subclasses of RGBABits.  Need support for direct conversion from RGBChar to GrayFloat and GrayDouble so that they use proper linear Y conversion.  Finally, some code is using Int2Char when it could load a word directly, and some code is incrementing pointers more than necessary.