More rigorous naming:  Renamed YVYUChar->UYVYChar and VYUYChar->YUYVChar to follow memory order convention.  Changed implementation of RGBAChar to follow memory order convention.
Changed get/setRGBA() to guarantee an order within machine word (which is of course invariant to endianness).

With the above two changes and a few other small tweaks to the code, it should now work for big-endian as well as small-endian, so removed endian warning.  However, this hasn't been tested on an actual big-endian yet.

Added RGBChar.  This is the format used by all the image i/o libraries that I currently link to.

Threw out instances of RGBABits, because they are misleadding and endian dependant.  RGBChar plus the rationalization of naming does most of what the bits instances were doing before.

Added == operator to GrayChar and GrayShort to recognize a RGBABits format that is compatible.

Still to do:  Make RGBAChar and RGBChar subclasses of RGBABits.