Add neighbors parameter to control finding local minima.
Use Gaussian information stored in Root objects associated with subsumed regions to avoid counting pixels more than once.  This implementation has a known memory leak.  Need to keep a list of detached Roots waiting for clean up.  Therefore, this is an "intermediate" checkin to preserve a milestone.  The next step is to rewrite all the supporting routines to share a data structure containing all important variables.  This could be the InterestMSER class itself, but then it wouldn't be thread safe, so will use a private structure in the implementation file.