Update revision history and add Sandia copyright notice.
Suppress min and max macros dragged in from somewhere by windows.h

Rearrange calls in Windows version of getTimestamp() so it reads the current counter before the frequency.  This is in keeping with the policy that we read the current time as close as possible to the start of the function.