Cache every image that results from some computation.  (Before, some of these were omitted.)
Compute the first couple of dogs at the next octave by decimation rather that a fresh difference.  Also, decimate the last image computed in the previous iteration and resume blurring there.  This saves a couple of convolutions, which are the most expensive part of the process.  (Working on a possible optimization which will place the downsampling at true octave bounaries, but it doesn't work yet.)