Use look up tables to convert gamma, rather than functions.  Hopefully this will make the process more efficient.
Changed 16 bit formats (GrayShort, RGBShort) to use gamma=1, on the assumption that the larger number of bits makes them more suitable for linear data, similar to float formats.  Not sure if images in the wild will satisfy this assumption.  Need more research.  Plan to add dynamically built luts for these formats at some point, so one can specify the gamma for the data.